pyproject.toml
[tool.poetry]
name = "dj_sinp_metadata"
version = "0.1.0"
description = "Django app to manage french SINP metadata"
authors = ["dbChiro <project@dbchiro.org>",]
license = "AGPLv3"
keywords = ["Django",]
readme = "README.rst"
homepage = "https://github.com/dbchiro/DjangoSinpMetadata"
repository = "https://github.com/dbchiro/DjangoSinpMetadata"
include = [
    "LICENSE",
]
packages = [
    { include = "sinp_metadata" },
]

[tool.poetry.dependencies]
python = ">=3.7,<4.0"
Django = "^3.2.5"
django-rest-framework = "^0.1.0"
dj-sinp-nomenclatures = "^0.1.1"
dj-dbchiro-roles = {path = "../dj_dbchiro_users"}
dj-sinp-organisms = {path = "../dj_sinp_organisms"}

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = {version = "^21.6b0", allow-prereleases = true}
Sphinx = "^4.1.1"
sphinx-rtd-theme = "^0.5.2"
rstcheck = "^3.3.1"
isort = "^5.9.2"
flake8 = "^3.9.2"
flake8-isort = "^4.0.0"
pre-commit = "^2.13.0"
python-decouple = "^3.4"
psycopg2-binary = "^2.9.1"
drf-yasg = "^1.20.0"

[tool.isort]
include_trailing_comma = true
line_length = 79
multi_line_output = 3
profile = "black"

[tool.black]
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | foo.py           # also separately exclude a file named foo.py in
                     # the root of the project
)
'''
include = '\.pyi?$'
line-length = 79
target-version = ['py37']



[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
